'Refactored by Sourcery'

+++ THE_ORGANIZER_3.2.py
@@ -46,21 +46,21 @@ def arrange_images():
             file for file in files if os.path.splitext(file)[1].lower() in img_ext
         ]
         print("\nSearching for Images", end="")
-        for i in range(10):
+        for _ in range(10):
             print(".", end="")
             sleep(0.2)
         print("Done")
-        if len(images) == 0:
+        if not images:
             print("No images found !!")
         else:
             print(f"Found {len(images)} images !!")
             print("\nSearching for 'Images' directory", end="")
-            for i in range(10):
+            for _ in range(10):
                 print(".", end="")
                 sleep(0.2)
             if os.path.exists("Images") == False:
                 print("Not Found !!\nSo creating", end="")
-                for i in range(10):
+                for _ in range(10):
                     print(".", end="")
                 sleep(0.2)
                 os.mkdir("Images")
@@ -80,21 +80,21 @@ def arrange_docs():
             file for file in files if os.path.splitext(file)[1].lower() in doc_ext
         ]
         print("\nSearching for Documents", end="")
-        for i in range(10):
+        for _ in range(10):
             print(".", end="")
             sleep(0.2)
         print("Done")
-        if len(documents) == 0:
+        if not documents:
             print("No documents found !!")
         else:
             print(f"Found {len(documents)} documents !!")
             print("\nSearching for 'Documents' directory", end="")
-            for i in range(10):
+            for _ in range(10):
                 print(".", end="")
                 sleep(0.2)
             if os.path.exists("Documents") == False:
                 print("Not Found !!\nSo creating", end="")
-                for i in range(10):
+                for _ in range(10):
                     print(".", end="")
                 sleep(0.2)
                 os.mkdir("Documents")
@@ -116,21 +116,21 @@ def arrange_videos():
             file for file in files if os.path.splitext(file)[1].lower() in video_ext
         ]
         print("\nSearching for Videos", end="")
-        for i in range(10):
+        for _ in range(10):
             print(".", end="")
             sleep(0.2)
         print("Done")
-        if len(videos) == 0:
+        if not videos:
             print("No videos found !!")
         else:
             print(f"Found {len(videos)} videos !!")
             print("\nSearching for 'Videos' directory", end="")
-            for i in range(10):
+            for _ in range(10):
                 print(".", end="")
                 sleep(0.2)
             if os.path.exists("Videos") == False:
                 print("Not Found !!\nSo creating", end="")
-                for i in range(10):
+                for _ in range(10):
                     print(".", end="")
                 sleep(0.2)
                 os.mkdir("Videos")
@@ -148,21 +148,21 @@ def arrange_audios():
             file for file in files if os.path.splitext(file)[1].lower() in audio_ext
         ]
         print("\nSearching for Audios", end="")
-        for i in range(10):
+        for _ in range(10):
             print(".", end="")
             sleep(0.2)
         print("Done")
-        if len(audios) == 0:
+        if not audios:
             print("No audio found !!")
         else:
             print(f"Found {len(audios)} audios !!")
             print("\nSearching for 'Audios' directory", end="")
-            for i in range(10):
+            for _ in range(10):
                 print(".", end="")
                 sleep(0.2)
             if os.path.exists("Audios") == False:
                 print("Not Found !!\nSo creating", end="")
-                for i in range(10):
+                for _ in range(10):
                     print(".", end="")
                 sleep(0.2)
                 os.mkdir("Audios")
@@ -192,21 +192,21 @@ def arrange_other():
             ):
                 others_ext.append(file)
         print("\nSearching for other files", end="")
-        for i in range(10):
+        for _ in range(10):
             print(".", end="")
             sleep(0.2)
         print("Done")
-        if len(others_ext) == 0:
+        if not others_ext:
             print("No others files found !!")
         else:
             print(f"Found {len(others_ext)} others files !!")
             print("\nSearching for 'Others' directory", end="")
-            for i in range(10):
+            for _ in range(10):
                 print(".", end="")
                 sleep(0.2)
             if os.path.exists("Others") == False:
                 print("Not Found !!\nSo creating", end="")
-                for i in range(10):
+                for _ in range(10):
                     print(".", end="")
                 sleep(0.2)
                 os.mkdir("Others")
@@ -227,17 +227,16 @@ def delete_empty_folder():
         empty_folders = []
         count = 0
         print("\nSearching for Empty folders", end="")
-        for i in range(10):
+        for _ in range(10):
             print(".", end="")
             sleep(0.3)
         for file in files:
-            if os.path.isdir(file) == True:
-                if len(os.listdir(file)) == 0:
-                    empty_folders.append(file)
-                    count += 1
-                    os.rmdir(file)
+            if os.path.isdir(file) == True and len(os.listdir(file)) == 0:
+                empty_folders.append(file)
+                count += 1
+                os.rmdir(file)
         print("Done !!")
-        if len(empty_folders) != 0:
+        if empty_folders:
             print(f"Successfully deleted {count} empty folders\n")
         else:
             print("No empty folders found !!\n")
